feat: updated ai-models tabs to extract common components and logic
